feat(registration): add select map flow

+++ app/src/main/java/com/tebet/mojual/data/models/Address.kt
@@ -6,8 +6,8 @@ class Address(
     var country: String? = null,
     var kecamatan: String? = null,
     var kelurahan: String? = null,
-    var latitude: Int? = null,
-    var longitude: Int? = null,
+    var latitude: Double? = null,
+    var longitude: Double? = null,
     var postalCode: String? = null,
     var rtrw: String? = null
 )
